requirements/base.txt
Django==1.9.6
Markdown==2.6.5
Pillow==3.1.0
Unidecode==0.04.18
Unipath==1.1
awesome-slugify==1.6.5
django-dotenv==1.4.1
django-filter==0.11.0
django-rest-auth==0.6.0
djangorestframework==3.3.2
psycopg2==2.6.1
six==1.10.0
